+++ src/app/app-routing.module.ts
@@ -6,8 +6,8 @@ import { InvoiceComponent } from './invoice/invoice.component';
 import {ReceivableSummaryComponent} from './receivable-summary/receivable-summary.component';
 import {AuthGuard} from './auth/auth.guard';
 import { ProfitlossComponent } from './profitloss/profitloss.component';
-import { AddExpensesComponent } from './add-expenses/add-expenses.component';
-import { ExpensesDetailComponent } from './expenses-detail/expenses-detail.component';
+import { AddExpensesComponent } from './add-travel/add-expenses.component';
+import { ExpensesDetailComponent } from './travel-details/expenses-detail.component';
 
 const routes: Routes = [
   { path: 'login', component: LoginComponent},
@@ -15,7 +15,7 @@ const routes: Routes = [
   { path: 'invoice', component: InvoiceComponent,  canActivate: [AuthGuard]},
   { path: 'summary', component: ReceivableSummaryComponent,  canActivate: [AuthGuard]},
   { path: 'PLStatus', component: ProfitlossComponent,  canActivate: [AuthGuard]},
-  { path: 'add-expenses', component: AddExpensesComponent,  canActivate: [AuthGuard]},
+  { path: 'add-travel', component: AddExpensesComponent,  canActivate: [AuthGuard]},
   { path: 'expense-details', component: ExpensesDetailComponent,  canActivate: [AuthGuard]},
   { path: '**', redirectTo: '/login', pathMatch: 'full' },
   { path: '', redirectTo: '/login', pathMatch: 'full' }
