+++ src/app/auth/auth.guard.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
-import {AuthService}  from '../auth.service';
+import {AuthService}  from '../services/auth.service';
 import { Routes, RouterModule } from '@angular/router';
 import { Router } from  "@angular/router";
