+++ src/app/home/home.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
-import {AuthService}  from '../auth.service';
+import {AuthService}  from '../services/auth.service';
 import { Router } from  "@angular/router";
-import {DataService} from '../data.service';
+import {DataService} from '../services/data.service';
 // Deploy on firebase
 //https://www.youtube.com/watch?v=mF7FTWHS3ys&t=13s
 // https://arth-e37f3.firebaseapp.com
