+++ src/app/invoice-table/invoice-table.component.html
@@ -52,7 +52,7 @@
     </tbody>
   </table>
    <div *ngIf="!hasDataToDisplay" class="text-center">
-    <label class="font-weight-bold">No data found</label> 
+    <label class="font-weight-normal">No data found</label> 
    </div>
    <div *ngIf="hasDataToDisplay" class="row">
       <div  class="text-center col-sm-12">
