+++ src/app/invoice/invoice.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, FormControl } from '@angular/forms';
-import {DataService} from '../data.service';
+import {DataService} from '../services/data.service';
 import { Router } from  "@angular/router";
 import { ValidatorFn, AbstractControl } from '@angular/forms';
 import {minValueValidator, greateThanZero}  from '../validators/validator';
