+++ src/app/login/login.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators, FormControl } from '@angular/forms';
-import {AuthService}  from '../auth.service';
+import {AuthService}  from '../services/auth.service';
 import { NgxUiLoaderService } from 'ngx-ui-loader'; 
 import { ToastrService } from 'ngx-toastr';
