+++ src/app/nav-panel/nav-panel.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from "@angular/router";
-import { AuthService } from "../auth.service";
+import { AuthService } from "../services/auth.service";
 
 @Component({
   selector: 'app-nav-panel',
