+++ src/app/profitloss/profitloss.component.html
@@ -89,7 +89,7 @@
   </table>
 
     <div *ngIf="!hasDataToDisplay" class="text-center col-12">
-      <label class="font-weight-bold">No data found</label> 
+      <label class="font-weight-normal">No data found</label> 
     </div>
 
 </div>
