+++ src/app/receivable-summary/receivable-summary.component.html
@@ -28,8 +28,8 @@
         <app-summary class="col-md-9 mt-3" [totValue]="totalValue" [noOfInv]="numberOfInvoices">
           </app-summary>
     </div>
-   
-    <h5 *ngIf="showDetailsFlag" class="text-center mt-3">Details of Invoices</h5>
+   <hr/>
+    <h5 *ngIf="showDetailsFlag" class="text-center mt-5 col-12">Details of Invoices</h5>
   
         <div class="row" *ngIf="showDetailsFlag">
             <table id="detailTable" class="table table-striped mt-3 mb-3 text-center col-md-12 col-sm-12" >
