+++ src/app/receivable-summary/receivable-summary.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { DataService} from '../data.service'
+import { DataService} from '../services/data.service';
  
 @Component({
   selector: 'app-receivable-summary',
