+++ src/app/utils/util.functions.ts
@@ -1,0 +1,13 @@
+
+export function multiFilter(array, filters) {
+    const filterKeys = Object.keys(filters);
+    // filters all elements passing the criteria
+    return array.filter((item) => {
+      // dynamically validate all filter criteria
+      return filterKeys.every(key => {
+        // ignores an empty filter
+        if (!filters[key].length) return true;
+        return filters[key].includes(item[key]);
+      });
+    });
+  }
